pyproject.toml
[project]
name = "kotone"
version = "1.0.4"
description = "A custom Discord app for Gakumas."
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "python-dotenv",
    "discord.py",
    "pyyaml",
    "flask",
    "google-api-python-client",
]

[project.urls]
Homepage = "https://github.com/ttineres/kotone-chan"

[project.scripts]
kotone-init = "kotone.bot:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["kotone"]
